Rephrase package purpose
The previous README strongly directed users towards `@recompile_invalidations`, which is discouraged for use in packages because it can hammer precompile times. I think the better use of this package is captured by "define the world of code before compilation." An indirect way to achieve the same goal is to push the number of methods above `max_methods`, even if the methods you're defining are spurious (though I don't mention this aspect in my edit).

I'm not certain this alternative phrasing really captures how this package is being used, and if so some rewording will be needed. I suspect such decisions are best left to those who know the ecosystem well, so do what you will with this PR.